Made restrictor and ballast commands csp admin tool compatible (#79)
* Made restrictor and ballast commands csp admin tool compatible

* Removed unnecessary comment

+++ AssettoServer/Server/EntryCar.cs
@@ -38,7 +38,7 @@ public partial class EntryCar : IEntryCar<ACTcpClient>
     public string Model { get; }
     public string Skin { get; }
     public int SpectatorMode { get; internal set; }
-    public int Ballast { get; internal set; }
+    public float Ballast { get; internal set; }
     public int Restrictor { get; internal set; }
     public List<ulong> AllowedGuids { get; internal set; } = new();
